additions

+++ routes/PostRoutes.ts
@@ -48,6 +48,22 @@ const PostRoutes = (app: Express) => {
         }
     })
 
+    app.put("/api/comment", async (req: Request, res: Response) => {
+        try {
+            const {comment, commentID} = req.body
+            if (!req.session.username || !comment || !commentID) return res.status(400).send("Bad request")
+            if (Number.isNaN(Number(commentID))) return res.status(400).send("Invalid commentID")
+            const badComment = functions.validateComment(comment as string) 
+            if (badComment) return res.status(400).send("Bad request")
+            const com = await sql.comment(Number(commentID))
+            if (com?.username !== req.session.username) return res.status(400).send("Bad request")
+            await sql.updateComment(Number(commentID), comment as string)
+            res.status(200).send("Success")
+        } catch {
+            res.status(400).send("Bad request") 
+        }
+    })
+
     app.delete("/api/post", async (req: Request, res: Response) => {
         try {
             const postID = req.query.postID
@@ -70,9 +86,52 @@ const PostRoutes = (app: Express) => {
     app.delete("/api/tag", async (req: Request, res: Response) => {
         try {
             const tag = req.query.tag as string
-            if (!tag) return res.status(400).send("Invalid postID")
-            if (!req.session.username) return res.status(400).send("Bad request")
+            if (!tag) return res.status(400).send("Invalid tag")
+            const tagExists = await sql.tag(tag.trim())
+            if (!req.session.username || !tagExists) return res.status(400).send("Bad request")
             await sql.deleteTag(tag.trim())
+            res.status(200).send("Success")
+        } catch {
+            res.status(400).send("Bad request") 
+        }
+    })
+
+    app.put("/api/tag", async (req: Request, res: Response) => {
+        try {
+            const {tag, key, description, image, aliases} = req.body
+            if (!req.session.username || !tag) return res.status(400).send("Bad request")
+            const tagObj = await sql.tag(tag)
+            if (!tagObj) return res.status(400).send("Bad request")
+            if (description) {
+                await sql.updateTag(tag, "description", description)
+            }
+            if (image) {
+                if (tagObj.image) {
+                    const imagePath = functions.getTagPath(tagObj.type, tagObj.image)
+                    await serverFunctions.deleteFile(imagePath)
+                }
+                const filename = `${tag}.${functions.fileExtension(image)}`
+                const imagePath = functions.getTagPath(tagObj.type, filename)
+                await serverFunctions.uploadFile(imagePath, Buffer.from(Object.values(image)))
+                await sql.updateTag(tag, "image", filename)
+                tagObj.image = filename
+            }
+            if (aliases?.[0]) {
+                await sql.purgeAliases(tag)
+                for (let i = 0; i < aliases.length; i++) {
+                    await sql.insertAlias(tag, aliases[i])
+                }
+            }
+            if (key.trim() !== tag) {
+                if (tagObj.image) {
+                    const newFilename = `${key.trim()}.${functions.fileExtension(image)}`
+                    const oldImagePath = functions.getTagPath(tagObj.type, tagObj.image)
+                    const newImagePath = functions.getTagPath(tagObj.type, newFilename)
+                    await serverFunctions.renameFile(oldImagePath, newImagePath)
+                    await sql.updateTag(tag, "image", newFilename)
+                }
+                await sql.updateTag(tag, "tag", key.trim())
+            }
             res.status(200).send("Success")
         } catch (e) {
             console.log(e)
@@ -84,16 +143,107 @@ const PostRoutes = (app: Express) => {
         try {
             const {postID, favorited} = req.body
             if (Number.isNaN(Number(postID))) return res.status(400).send("Invalid postID")
-            if (!favorited || !req.session.username) return res.status(400).send("Bad request")
+            if (favorited == null || !req.session.username) return res.status(400).send("Bad request")
+            const favorite = await sql.favorite(Number(postID), req.session.username)
             if (favorited) {
-                await sql.insertFavorite(Number(postID), req.session.username)
+                if (!favorite) await sql.insertFavorite(Number(postID), req.session.username)
             } else {
-                const favorite = await sql.favorite(Number(postID), req.session.username)
                 if (favorite) await sql.deleteFavorite(favorite.favoriteID)
+            }
+            res.status(200).send("Success")
+        } catch {
+            res.status(400).send("Bad request") 
+        }
+    })
+
+    app.get("/api/favorite", async (req: Request, res: Response) => {
+        try {
+            const postID = req.query.postID
+            if (Number.isNaN(Number(postID))) return res.status(400).send("Invalid postID")
+            if (!req.session.username) return res.status(400).send("Bad request")
+            const favorite = await sql.favorite(Number(postID), req.session.username)
+            res.status(200).send(favorite)
+        } catch {
+            res.status(400).send("Bad request") 
+        }
+    })
+
+    app.get("/api/favorites", async (req: Request, res: Response) => {
+        try {
+            if (!req.session.username) return res.status(400).send("Bad request")
+            const favorites = await sql.favorites(req.session.username)
+            res.status(200).send(favorites)
+        } catch {
+            res.status(400).send("Bad request") 
+        }
+    })
+
+    app.get("/api/uploads", async (req: Request, res: Response) => {
+        try {
+            if (!req.session.username) return res.status(400).send("Bad request")
+            const uploads = await sql.uploads(req.session.username)
+            res.status(200).send(uploads)
+        } catch {
+            res.status(400).send("Bad request") 
+        }
+    })
+
+    app.post("/api/cuteness", async (req: Request, res: Response) => {
+        try {
+            const {postID, cuteness} = req.body
+            if (Number.isNaN(Number(postID)) || Number.isNaN(Number(cuteness))) return res.status(400).send("Bad request")
+            if (!req.session.username) return res.status(400).send("Bad request")
+            if (Number(cuteness) < 0 || Number(cuteness) > 1000) return res.status(400).send("Bad request")
+            const cute = await sql.cuteness(Number(postID), req.session.username)
+            if (cute) {
+                await sql.updateCuteness(Number(postID), req.session.username, Number(cuteness))
+            } else {
+                await sql.insertCuteness(Number(postID), req.session.username, Number(cuteness))
             }
             res.status(200).send("Success")
         } catch (e) {
             console.log(e)
+            res.status(400).send("Bad request") 
+        }
+    })
+
+    app.get("/api/cuteness", async (req: Request, res: Response) => {
+        try {
+            const postID = req.query.postID
+            if (Number.isNaN(Number(postID))) return res.status(400).send("Invalid postID")
+            if (!req.session.username) return res.status(400).send("Bad request")
+            const cuteness = await sql.cuteness(Number(postID), req.session.username)
+            res.status(200).send(cuteness)
+        } catch {
+            res.status(400).send("Bad request") 
+        }
+    })
+
+    app.delete("/api/cuteness", async (req: Request, res: Response) => {
+        try {
+            const postID = req.query.postID
+            if (Number.isNaN(Number(postID))) return res.status(400).send("Invalid postID")
+            if (!req.session.username) return res.status(400).send("Bad request")
+            const cuteness = await sql.cuteness(Number(postID), req.session.username)
+            if (!cuteness) return res.status(400).send("Bad request")
+            await sql.deleteCuteness(cuteness.cutenessID)
+            res.status(200).send(cuteness)
+        } catch {
+            res.status(400).send("Bad request") 
+        }
+    })
+
+    app.post("/api/aliastag", async (req: Request, res: Response) => {
+        try {
+            const {tag, aliasTo} = req.body
+            if (!req.session.username || !tag || !aliasTo) return res.status(400).send("Bad request")
+            const exists = await sql.tag(aliasTo)
+            if (!exists) return res.status(400).send("Bad request")
+            await sql.renameTagMap(tag, aliasTo)
+            await sql.deleteTag(tag)
+            await sql.insertAlias(aliasTo, tag)
+            res.status(200).send("Success")
+        } catch {
             res.status(400).send("Bad request") 
         }
     })
